# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VendlySys - SaaS multiempresa para salões e clínicas
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /financeiro:
    get:
      operationId: listLancamentos
      tags: [financeiro]
      summary: Lista lançamentos financeiros
      parameters:
        - name: data_inicio
          in: query
          schema:
            type: string
            format: date
        - name: data_fim
          in: query
          schema:
            type: string
            format: date
        - name: status_pagamento
          in: query
          schema:
            type: string
            enum: [pendente, pago, cancelado]
        - name: forma_pagamento
          in: query
          schema:
            type: string
            enum: [dinheiro, pix, cartao_credito, cartao_debito, transferencia]
        - name: profissional_id
          in: query
          schema:
            type: integer
        - name: tipo
          in: query
          schema:
            type: string
            enum: [receita, despesa]
      responses:
        "200":
          description: Lista de lançamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lancamento"
  /financeiro/{id}:
    put:
      operationId: updateLancamento
      tags: [financeiro]
      summary: Atualiza lançamento financeiro
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLancamentoBody"
      responses:
        "200":
          description: Lançamento atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lancamento"
  /financeiro/{id}/pagar:
    post:
      operationId: marcarPago
      tags: [financeiro]
      summary: Marca lançamento como pago
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarcarPagoBody"
      responses:
        "200":
          description: Marcado como pago
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lancamento"
...
  /caixa/fechamento:
    get:
      operationId: getFechamentoCaixa
      tags: [financeiro]
      summary: Resumo de fechamento de caixa do período
      parameters:
        - name: data_inicio
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: data_fim
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Resumo do caixa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FechamentoCaixa"
...
components:
  schemas:
...
    Lancamento:
      type: object
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        agendamento_id:
          type: integer
        cliente_nome:
          type: string
        profissional_nome:
          type: string
        servico:
          type: string
        data:
          type: string
          format: date
        valor:
          type: number
        valor_pago:
          type: number
        desconto:
          type: number
        status_pagamento:
          type: string
          enum: [pendente, pago, cancelado]
        forma_pagamento:
          type: string
          enum: [dinheiro, pix, cartao_credito, cartao_debito, transferencia]
        tipo:
          type: string
          enum: [receita, despesa]
        observacoes:
          type: string
      required: [id, empresa_id, servico, data, valor, status_pagamento, tipo]
    UpdateLancamentoBody:
      type: object
      properties:
        forma_pagamento:
          type: string
          enum: [dinheiro, pix, cartao_credito, cartao_debito, transferencia]
        desconto:
          type: number
        observacoes:
          type: string
    MarcarPagoBody:
      type: object
      properties:
        forma_pagamento:
          type: string
          enum: [dinheiro, pix, cartao_credito, cartao_debito, transferencia]
        valor_pago:
          type: number
        observacoes:
          type: string
      required: [forma_pagamento]
...
    FechamentoCaixa:
      type: object
      properties:
        data_inicio:
          type: string
          format: date
        data_fim:
          type: string
          format: date
        total_receitas:
          type: number
        total_despesas:
          type: number
        total_recebido:
          type: number
        total_pendente:
          type: number
        saldo_periodo:
          type: number
        qtd_atendimentos:
          type: integer
        ticket_medio:
          type: number
        por_forma_pagamento:
          type: array
          items:
            type: object
            properties:
              forma:
                type: string
              total:
                type: number
              quantidade:
                type: integer
            required: [forma, total, quantidade]
      required: [data_inicio, data_fim, total_receitas, total_despesas, total_recebido, total_pendente, saldo_periodo, qtd_atendimentos, ticket_medio, por_forma_pagamento]
...